@@ -15,48 +15,44 @@ dependencies:
     sdk: flutter
   dartz: ^0.10.1
   enum_to_string: ^2.0.1
-  file: ^6.1.2
+  file: ^7.0.0
   cross_file: ^0.3.2
-  flutter_archive: ^4.1.0
-  freezed_annotation: ^1.1.0
-  internet_connection_checker: ^0.0.1+3
-  json_annotation: ^4.4.0
-  drift: ^1.2.0
+  flutter_archive: ^6.0.0
+  freezed_annotation: ^2.4.1
+  internet_connection_checker: ^1.0.0+1
+  json_annotation: ^4.8.1
+  drift: ^2.16.0
   sqlite3_flutter_libs: ^0.5.2
   path: ^1.8.1
   path_provider: ^2.0.7
-  moor_db_viewer: ^5.0.0
+  drift_db_viewer: ^2.0.0
   android_path_provider: ^0.3.0
-  reactive_forms: ^11.0.0
-  hooks_riverpod: ^1.0.3
-  reactive_multi_image_picker:
-    git:
-      url: https://github.com/victorcatn/reactive_multi_image_picker.git
-      ref: multi_platform_support
+  reactive_forms: ^16.1.1
+  hooks_riverpod: ^ 2.5.1
+  #reactive_multi_image_picker:
+    #git:
+      #url: https://github.com/victorcatn/reactive_multi_image_picker.git
+      #ref: multi_platform_support
   rflutter_alert: ^2.0.4
-  sentry_flutter: ^6.2.2
+  sentry_flutter: ^7.18.0
   shared_preferences: ^2.0.11
   simple_tooltip: ^1.0.0
   url_launcher: ^6.0.17
-  go_router: ^2.5.7
-  connectivity_plus: ^2.2.0
-  textfield_tags:
-    # git:
-    #   url: https://github.com/victorcatn/textfield_tags.git
-    path: ../textfield_tags
-  uuid: ^3.0.5
+  go_router: ^13.2.1
+  connectivity_plus: ^6.0.1
+  uuid: ^4.3.3
   rxdart: ^0.27.3
   diacritic: ^0.1.3
 
 dev_dependencies:
-  build_runner: ^2.1.7
-  drift_dev: ^1.2.0
-  flutter_launcher_icons: ^0.9.2
-  freezed: ^1.1.0
-  json_serializable: ^6.1.3
-  flutter_lints: ^1.0.4
-  test: ^1.20.1
-  mockito: ^5.0.17
+  build_runner: ^2.4.8
+  drift_dev: ^2.16.0
+  flutter_launcher_icons: ^0.13.1
+  freezed: ^2.4.7
+  json_serializable: ^6.7.1
+  flutter_lints: ^3.0.1
+  test: ^1.25.2
+  mockito: ^5.4.4
   file_testing: ^3.0.0
   
 
