@@ -44,42 +44,56 @@ import 'creacion_form_controller.dart';
 
 /// Marca como requerido el textField de 'nuevoTipoDeInspeccion',
 /// en caso de que se haya seleccionado 'Otra' como tipo de inspeccion.
-ValidatorFunction nuevoTipoDeInspeccionValidator(
-  FormControl<String?> tipoDeInspeccionControl,
-  FormControl<String?> nuevoTipoDeInspeccionControl,
-) =>
-    (AbstractControl<dynamic> control) {
-      final ti = tipoDeInspeccionControl.value;
-      final nti = nuevoTipoDeInspeccionControl.value;
-      if (ti == null || nti == null) return null;
 
-      final error = {ValidationMessage.required: true};
+class NuevoTipoDeInspeccionValidator extends Validator<dynamic> {
+  final FormControl<String?> tipoDeInspeccionControl;
+  final FormControl<String?> nuevoTipoDeInspeccionControl;
 
-      if (ti == CreacionFormController.otroTipoDeInspeccion &&
-          nti.trim().isEmpty) {
-        nuevoTipoDeInspeccionControl.setErrors(error);
-        //nuevoTipoDeinspeccion.markAsTouched();
-      } else {
-        nuevoTipoDeInspeccionControl.removeError(ValidationMessage.required);
-      }
-      return null;
-    };
+  NuevoTipoDeInspeccionValidator(
+    this.tipoDeInspeccionControl,
+    this.nuevoTipoDeInspeccionControl,
+  );
+
+  @override
+  Map<String, dynamic>? validate(AbstractControl<dynamic> control) {
+    final ti = tipoDeInspeccionControl.value;
+    final nti = nuevoTipoDeInspeccionControl.value;
+    if (ti == null || nti == null) return null;
+
+    final error = {ValidationMessage.required: true};
+
+    if (ti == CreacionFormController.otroTipoDeInspeccion &&
+        nti.trim().isEmpty) {
+      nuevoTipoDeInspeccionControl.setErrors(error);
+      //nuevoTipoDeinspeccion.markAsTouched();
+    } else {
+      nuevoTipoDeInspeccionControl.removeError(ValidationMessage.required);
+    }
+    return null;
+  }
+}
 
 /// Validación que verifica que el valor del textfield para 'minimo' sea menor que el valor introducido en el textfield 'maximo'.
 /// En las preguntas numéricas
-ValidatorFunction verificarRango(
-        {required FormControl<double> controlMinimo,
-        required FormControl<double> controlMaximo}) =>
-    (AbstractControl<dynamic> _) {
-      if (controlMinimo.value == null || controlMaximo.value == null) {
-        controlMaximo.removeError('verificarRango');
-        return null;
-      }
-      if (controlMinimo.value! > controlMaximo.value!) {
-        controlMaximo.setErrors({'verificarRango': true});
-      } else {
-        controlMaximo.removeError('verificarRango');
-      }
 
+class VerificarRango extends Validator<dynamic> {
+  final FormControl<double> controlMinimo;
+  final FormControl<double> controlMaximo;
+
+  VerificarRango(this.controlMinimo, this.controlMaximo);
+
+  @override
+  Map<String, dynamic>? validate(AbstractControl<dynamic> control) {
+    if (controlMinimo.value == null || controlMaximo.value == null) {
+      controlMaximo.removeError('verificarRango');
       return null;
-    };
+    }
+    if (controlMinimo.value! > controlMaximo.value!) {
+      controlMaximo.setErrors({'verificarRango': true});
+    } else {
+      controlMaximo.removeError('verificarRango');
+    }
+
+    return null;
+  }
+}
