@@ -7,18 +7,18 @@ import 'inspecciones_remote_repository.dart';
 import 'organizacion_repository.dart';
 import 'user_repository.dart';
 
-final appRepositoryProvider = Provider((ref) => AppRepository(ref.read));
+final appRepositoryProvider = Provider((ref) => AppRepository(ref));
 
-final userRepositoryProvider = Provider((ref) => UserRepository(ref.read));
+final userRepositoryProvider = Provider((ref) => UserRepository(ref));
 
 final cuestionariosRepositoryProvider =
-    Provider((ref) => CuestionariosRepository(ref.read));
+    Provider((ref) => CuestionariosRepository(ref));
 
 final inspeccionesRemoteRepositoryProvider =
-    Provider((ref) => InspeccionesRemoteRepository(ref.read));
+    Provider((ref) => InspeccionesRemoteRepository(ref));
 
 final organizacionRepositoryProvider =
-    Provider((ref) => OrganizacionRepository(ref.read));
+    Provider((ref) => OrganizacionRepository(ref));
 
 final fileSystemProvider = Provider<FileSystem>(
     (ref) => throw Exception("no se ha definido el FileSystem"));
