@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:inspecciones/infrastructure/drift_database.dart';
-import 'package:moor_db_viewer/moor_db_viewer.dart';
 
 class InspeccionesDbViewerPage extends ConsumerWidget {
   const InspeccionesDbViewerPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context, ref) {
-    return MoorDbViewer(ref.read(driftDatabaseProvider));
+    //TODO: revisar
+    return const Text(
+        'Hola'); /* MoorDbViewer(ref.read(driftDatabaseProvider)); */
   }
 }
